requirements-airflow.txt
pandas
psycopg2-binary
python-dotenv
dbt-core
dbt-postgres
numpy
scikit-learn
